+++ src/Compiler/Driver/ParseAndCheckInputs.fs
@@ -625,7 +625,8 @@ let ParseOneInputLexbuf (tcConfig: TcConfig, lexResourceManager, lexbuf, fileNam
                     )
 
                 // Report the statistics for testing purposes
-                if tcConfig.reportNumDecls then ReportParsingStatistics res
+                if tcConfig.reportNumDecls then
+                    ReportParsingStatistics res
 
                 res)
 
@@ -1075,7 +1076,7 @@ let GetInitialTcState (m, ccuName, tcConfig: TcConfig, tcGlobals, tcImports: TcI
 
     // Create a ccu to hold all the results of compilation
     let ccuContents =
-        Construct.NewCcuContents ILScopeRef.Local m ccuName (Construct.NewEmptyModuleOrNamespaceType Namespace)
+        Construct.NewCcuContents ILScopeRef.Local m ccuName (Construct.NewEmptyModuleOrNamespaceType(Namespace true))
 
     let ccuData: CcuData =
         {
@@ -1113,7 +1114,7 @@ let GetInitialTcState (m, ccuName, tcConfig: TcConfig, tcGlobals, tcImports: TcI
         tcsCreatesGeneratedProvidedTypes = false
         tcsRootSigs = Zmap.empty qnameOrder
         tcsRootImpls = Zset.empty qnameOrder
-        tcsCcuSig = Construct.NewEmptyModuleOrNamespaceType Namespace
+        tcsCcuSig = Construct.NewEmptyModuleOrNamespaceType(Namespace true)
         tcsImplicitOpenDeclarations = openDecls0
     }
 
@@ -1219,7 +1220,7 @@ let CheckOneInput
                 let typeCheckOne =
                     if skipImplIfSigExists && hadSig then
                         (EmptyTopAttrs, CreateEmptyDummyImplFile qualNameOfFile rootSigOpt.Value, Unchecked.defaultof<_>, tcImplEnv, false)
-                        |> Cancellable.ret
+                        |> cancellable.Return
                     else
                         CheckOneImplFile(
                             tcGlobals,