+++ src/Compiler/Optimize/InnerLambdasToTopLevelFuncs.fs
@@ -86,8 +86,8 @@ let isDelayedRepr (f: Val) e =
 
 // REVIEW: these should just be replaced by direct calls to mkLocal, mkCompGenLocal etc.
 // REVIEW: However these set an arity whereas the others don't
-let mkLocalNameTypeArity compgen m name ty topValInfo =
-    Construct.NewVal(name, m, None, ty, Immutable, compgen, topValInfo, taccessPublic, ValNotInRecScope, None, NormalVal, [], ValInline.Optional, XmlDoc.Empty, false, false, false, false, false, false, None, ParentNone)
+let mkLocalNameTypeArity compgen m name ty valReprInfo =
+    Construct.NewVal(name, m, None, ty, Immutable, compgen, valReprInfo, taccessPublic, ValNotInRecScope, None, NormalVal, [], ValInline.Optional, XmlDoc.Empty, false, false, false, false, false, false, None, ParentNone)
 
 //-------------------------------------------------------------------------
 // definitions: TLR, arity, arity-met, arity-short
@@ -771,8 +771,6 @@ let FlatEnvPacks g fclassM topValS declist (reqdItemsMap: Zmap<BindingGroupShari
 
        // dump
        if verboseTLR then
-           let bindingL bind = bindingL g bind
-
            dprintf "tlr: packEnv envVals =%s\n" (showL (listL valL env.ReqdVals))
            dprintf "tlr: packEnv envSubs =%s\n" (showL (listL valL env.ReqdSubEnvs))
            dprintf "tlr: packEnv vals =%s\n" (showL (listL valL vals))
@@ -795,18 +793,6 @@ let FlatEnvPacks g fclassM topValS declist (reqdItemsMap: Zmap<BindingGroupShari
 // step3: chooseEnvPacks
 //-------------------------------------------------------------------------
 
-#if DEBUG
-let DumpEnvPackM g envPackM =
-    let bindingL bind = bindingL g bind
-    for KeyValue(fc, packedReqdItems) in envPackM do
-        dprintf "packedReqdItems: fc = %A\n" fc
-        dprintf "         reqdTypars = %s\n" (showL (commaListL (List.map typarL packedReqdItems.ep_etps)))
-        dprintf "         aenvs = %s\n" (showL (commaListL (List.map valL packedReqdItems.ep_aenvs)))
-        dprintf "         pack = %s\n" (showL (semiListL (List.map bindingL packedReqdItems.ep_pack)))
-        dprintf "         unpack = %s\n" (showL (semiListL (List.map bindingL packedReqdItems.ep_unpack)))
-        dprintf "\n"
-#endif
-
 /// For each fclass, have an env.
 /// Required to choose an PackedReqdItems,
 /// e.g. deciding whether to tuple up the environment or not.
@@ -818,9 +804,6 @@ let DumpEnvPackM g envPackM =
 let ChooseReqdItemPackings g fclassM topValS  declist reqdItemsMap =
     if verboseTLR then dprintf "ChooseReqdItemPackings------\n"
     let envPackM = FlatEnvPacks g fclassM topValS  declist reqdItemsMap
-#if DEBUG
-    if verboseTLR then DumpEnvPackM g envPackM
-#endif
     envPackM
 
 //-------------------------------------------------------------------------